fb-autoposter/requirements.txt
anthropic>=0.40.0
pydantic>=2.0
python-dotenv>=1.0
opencv-python>=4.8
discord.py>=2.3
